Thylacine Blendspace and montage
Nothing final yet. Just some tests that are currently not working but in a good path.
Some renaming inside blueprints to make things clearer

Focus on: Modify bone, roar before chasing the player, check the random points in a radius not to fall close to each other, notifiers on montage.